+++ src/main.cpp
@@ -1,3 +1,5 @@
+#include <Geode/Geode.hpp>
+#include <Geode/ui/Layout.hpp>
 #include <Geode/modify/GJGarageLayer.hpp>
 #include <Geode/modify/LevelInfoLayer.hpp>
 #include "layers/SogLayer.hpp"
@@ -21,6 +23,9 @@ struct GJGarageLayerHook : Modify<GJGarageLayerHook, GJGarageLayer> {
 		sogBtn->setPositionY(-2);
 		sogMenu->addChild(sogBtn);
 
+		sogBtn->setID("soggy-button"_spr);
+		sogMenu->setID("soggy-menu"_spr);
+
 		sogMenu->setPosition(ccp((winSize.width / 2) + 141, director->getScreenTop() - 23));
 
 		this->addChild(sogMenu);
@@ -32,28 +37,37 @@ struct GJGarageLayerHook : Modify<GJGarageLayerHook, GJGarageLayer> {
 	}
 };
 
-// struct LevelInfoLayerHook : Modify<LevelInfoLayerHook, LevelInfoLayer> {
-// 	bool init(GJGameLevel* level, bool challenge) {
-// 		if(!LevelInfoLayer::init(level, challenge)) {
-// 			return false;
-// 		}
+struct LevelInfoLayerHook : Modify<LevelInfoLayerHook, LevelInfoLayer> {
+	bool init(GJGameLevel* level, bool challenge) {
+		if(!LevelInfoLayer::init(level, challenge)) {
+			return false;
+		}
 
-// 		auto sogSpr = CCSprite::createWithSpriteFrameName("GJ_soggyRope_001.png"_spr);
-// 		auto sogBtn = CCMenuItemSpriteExtra::create(sogSpr, this, menu_selector(LevelInfoLayerHook::onSoggy));
-// 		auto sogMenu = CCMenu::create();
+		auto sogSpr = CCSprite::createWithSpriteFrameName("GJ_soggyRope_001.png"_spr);
+		auto sogBtn = CCMenuItemSpriteExtra::create(sogSpr, this, menu_selector(LevelInfoLayerHook::onSoggy));
+		
+		auto garageMenu = static_cast<CCMenu*>(getChildByID("garage-menu"));
+		if(!garageMenu) {
+			log::error("LevelInfoLayer 'garage-menu' does not exist, returning.");
+			return true;
+		}
 
-// 		sogBtn->m_animationType = MenuAnimationType::Move;
-// 		sogBtn->m_startPosition = sogSpr->getPosition();
-// 		sogBtn->m_duration = 0.2f;
-// 		sogBtn->m_unselectedDuration = 0.2f;
-// 		sogBtn->m_offset = ccp(0, -8.f);
-// 		sogMenu->addChild(sogBtn);
+		garageMenu->setLayout(AxisLayout::create());
+		garageMenu->setPositionX(garageMenu->getPositionX() - 25);
 
-// 		addChild(sogMenu, 100);
+		sogBtn->m_animationType = MenuAnimationType::Move;
+		sogBtn->m_startPosition = sogSpr->getPosition();
+		sogBtn->m_duration = 0.2f;
+		sogBtn->m_unselectedDuration = 0.2f;
+		sogBtn->m_offset = ccp(0, -8.f);
+		sogBtn->setID("soggy-button"_spr);
 
-// 		return true;
-// 	}
-// 	void onSoggy(CCObject*) {
-// 		CCDirector::sharedDirector()->pushScene(CCTransitionMoveInT::create(0.5f, SogLayer::scene(true)));
-// 	}
-// };
+		garageMenu->addChild(sogBtn);
+		garageMenu->updateLayout();
+
+		return true;
+	}
+	void onSoggy(CCObject*) {
+		CCDirector::sharedDirector()->pushScene(CCTransitionMoveInT::create(0.5f, SogLayer::scene(true)));
+	}
+};